pyproject.toml
[project]
name = "youtube-video-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "patool>=4.0.0",
    "pyinstaller>=6.13.0",
    "pyqt6>=6.9.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "yt-dlp>=2025.3.31",
]
